requirements.txt
requests==2.32.3
beautifulsoup4==4.12.3
readability-lxml==0.8.1
lxml_html_clean==0.2.2
psycopg2-binary==2.9.9
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
